rock sample detection implemented
